SCS2/requirements.txt
cryptography==38.0.3
Django==4.1.3
gspread==5.6.2
oauth2client==4.1.3
PyGithub==1.56
secrets==1.0.2
